Update orderForm.html
Changed CSS to not use pixels but use rem, it is not advised to use pixels for measurements because it will not scale with the web page. It is better to use rem so it will scale automatically based on screen size. Also added some more margin and padding to tidy up the page.